requirements.txt
simplegmail==4.1.1
langchain_community==0.0.21
langchain-openai==0.0.6
langchainhub==0.1.14
langchain==0.1.8
langgraph==0.0.24
langchain-together==0.0.2.post1
langchain-experimental==0.0.52
pinecone-client==3.0.3
pdf2image==1.17.0
pytesseract==0.3.10
tiktoken==0.6.0
Flask==3.0.2
Flask-Cors==4.0.0
cohere==4.48
flask-swagger==0.2.14
schedule==1.2.1
flasgger==0.9.7.1
gunicorn==22.0.0
pyautogen==0.2.26
slack-sdk==3.27.1
slackeventsapi==3.0.1